refactor HandleGetTranscodeDecision
